Updated time daemon
